feat: allow enum parameters to be specified as their string literal equivalents
Expand the types for `DecryptionAlgorithm` and `Options` (used to
create the `FingerprintJsServerApiClient`) to allow the enum parameters
to also be specified as their string literal equivalents. This allows
TypeScript users of the SDK to use the string literal equivalents
without having to be aware of the enums.

To achieve this, use ["Template Literal Types"](https://www.typescriptlang.org/docs/handbook/2/template-literal-types.html)
to expand the values for string-valued enums to a union type without having to
duplicate the string constants.

This approach results in no differences to the JS generated
by TypeScript compilation, and as a result, this change just results in
a backwards-compatible change to the TypeScript types.

Also:
- Add unit tests to ensure that using the string literals does not
  result in compilation errors.
- Add a unit test to test usage of the API key as a query parameter